Iteration 1: Implement a Singleton
